+++ crates/pdk-api/src/api.rs
@@ -54,7 +54,7 @@ api_struct!(
         /// Override the tool inventory directory (where all versions are installed).
         /// This is an advanced feature and should only be used when absolutely necessary.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub override_dir: Option<PathBuf>,
+        pub override_dir: Option<VirtualPath>,
 
         /// Suffix to append to all versions when labeling directories.
         #[serde(skip_serializing_if = "Option::is_none")]